requirements_ai_training.txt
accelerate==1.11.0
facenet-pytorch==2.5.3
lightning==2.5.5
lightning-utilities==0.15.2
mediapipe==0.10.21
numpy==1.26.4
open_clip_torch==3.2.0
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
pandas==2.3.3
pillow==11.3.0
pyarrow==21.0.0
pytorch-lightning==2.5.5
scipy==1.16.2
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.2
torchreid @ git+https://github.com/KaiyangZhou/deep-person-reid.git@566a56a2cb255f59ba75aa817032621784df546a
torchvision==0.23.0
tqdm==4.67.1
